+++ website-api/dashboard.js
@@ -7,25 +7,7 @@ const {
     getUserPlans
 } = require('../database/plans.js');
 
-// Auth middleware to verify token and extract userId
-const authMiddleware = (req, res, next) => {
-    const sessionToken = req.cookies.session;
-    console.log('sessionToken:', sessionToken);
-    if (!sessionToken) {
-        console.log('No session token');
-        return res.status(401).json({ message: 'Not authenticated' });
-    }
-    
-    try {
-        const decoded = jwt.verify(sessionToken, process.env.JWT_SECRET);
-        console.log('decoded:', decoded);
-        req.userId = decoded.userId; // Attach userId to request
-        next();
-    } catch (err) {
-        console.error('Invalid token:', err);
-        return res.status(401).json({ message: 'Invalid token' });
-    }
-};
+const { authMiddleware } = require('./middleware.js');
 
 // Get all plans for a user
 router.get('/user-plans', authMiddleware, async (req, res) => {
